Update code for upcoming treecontrols component
